class03-20250427/04_openrouter/openrouter/pyproject.toml
[project]
name = "openrouter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hassan Ali Khan", email = "73381969+hassan-ak@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.13",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[project.scripts]
openrouter_gemini = "openrouter.openrouter_directly:openrouter_gemini"
openrouter_deepseek = "openrouter.openrouter_directly:openrouter_deepseek"
agent_gemini = "openrouter.agent:agent_gemini"
agent_deepseek = "openrouter.agent:agent_deepseek"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
